backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clipjot"
version = "0.1.0"
description = "ClipJot bookmark management backend"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.4.0",
    "fastlite>=0.0.9",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
clipjot = "app.cli:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
